+++ test/support/alternate_user.ex
@@ -1,0 +1,3 @@
+defmodule Mapail.AlternateUser do
+  defstruct [:first_name, :username, :password, :email]
+end