+++ test/support/user.ex
@@ -1,0 +1,3 @@
+defmodule Mapail.User do
+  defstruct [:first_name, :username, :password]
+end